+++ src/components/DatePicker/styles.ts
@@ -1,329 +1,364 @@
-import styled from 'styled-components';
-import { CalendarContainer } from 'react-datepicker';
+import styled, { css, FlattenSimpleInterpolation } from 'styled-components';
+import colors from '../../styles/colors';
 import CustomButton from '../CustomButton';
-import { farmerConnectTheme } from '../Theme';
-import { Input as DefaultInput } from '../Input/styles';
-export { HelperText } from '../Input/styles';
+import Calendar from '../Icons/Calendar';
+import CustomSelect from '../Select';
+import SmallSelect from '../SmallSelect';
+import Typography from '../Typography';
+import { Variants } from './interfaces';
 
-export const DatePickerWrapper = styled.div`
-  position: relative;
-
-  .react-datepicker-popper {
-    z-index: 50;
-  }
-
-  > svg {
-    color: ${farmerConnectTheme.colors.fc_black_70};
-    position: absolute;
-    right: 0.5rem;
-    top: 50%;
-    transform: translateY(-50%);
-  }
-`;
-
-type DatePickerInputProps = {
-  error?: boolean;
+const variants: Record<Variants, FlattenSimpleInterpolation> = {
+  default: css`
+    --input-bg: ${colors.fc_beige};
+    --input-box-shadow: 0.125rem 0.125rem 0 ${colors.fc_green};
+    --input-min-height: 2.5rem;
+    --input-border-color: transparent;
+    --input-font-size: 0.875rem;
+    --input-font-weight: 500;
+    --input-icon-height: 1.5rem;
+    --input-left-padding: 1rem;
+    --input-border-color-error: transparent;
+    --input-box-shadow-error: 0.125rem 0.125rem 0 ${colors.fc_red};
+    --mask-divider-color: ${colors.fc_beige};
+    --range-bg-color: ${colors.fc_black_10};
+    --range-font-color: ${colors.fc_black_100};
+  `,
+  small: css`
+    --input-bg: ${colors.fc_white};
+    --input-box-shadow: none;
+    --input-min-height: 2rem;
+    --input-border-color: ${colors.fc_black_100};
+    --input-font-size: 0.6875rem;
+    --input-font-weight: 700;
+    --input-icon-height: 1.125rem;
+    --input-left-padding: 0.5rem;
+    --input-border-color-error: ${colors.fc_red};
+    --input-box-shadow-error: none;
+    --mask-divider-color: ${colors.fc_white};
+    --range-bg-color: ${colors.fc_black_5};
+    --range-font-color: ${colors.fc_black_100};
+  `,
 };
 
-export const DatePickerSmallInput = styled.input<DatePickerInputProps>`
-  font-family: 'Red Hat Text', sans-serif;
-  cursor: default;
-  width: 100%;
-  box-sizing: border-box;
-  border: 1px solid ${({ error }) => (error ? farmerConnectTheme.colors.fc_red : farmerConnectTheme.colors.fc_black_70)};
-  border-radius: 0.75rem;
-  outline: none;
-  font-size: 0.6875rem;
-  color: ${farmerConnectTheme.colors.fc_black_100};
-  font-weight: 700;
-  line-height: 1rem;
-  padding: 0.5rem 1.75rem 0.5rem 0.75rem;
-  height: 2rem;
-  text-transform: uppercase;
-  background-color: transparent;
-
-  &::placeholder {
-    color: ${farmerConnectTheme.colors.fc_black_30};
-    font-weight: 400;
-    text-transform: none;
-  }
-`;
-
-export const DatePickerDefaultInput = styled(DefaultInput)<DatePickerInputProps>`
-  text-transform: uppercase;
-  &::placeholder {
-    text-transform: none;
-  }
-`;
-
-export const Calendar = styled(CalendarContainer)<{ monthsShown: number }>`
-  box-sizing: border-box;
-  border-radius: 0.75rem;
-  box-shadow: 0 0 0 1px ${farmerConnectTheme.colors.fc_black_100};
+type ContainerProps = {
+  variant: Variants;
+};
+export const Container = styled.div<ContainerProps>`
   position: relative;
-  padding: 5rem 1rem 1rem 1rem;
-  display: grid;
-  grid-template-columns: repeat(${(props) => props.monthsShown}, 1fr);
-  gap: 0 2rem;
-  background-color: ${farmerConnectTheme.colors.fc_white};
+  font-family: 'Red Hat Text', sans-serif;
+  ${(props) => variants[props.variant]}
+`;
 
-  .react-datepicker {
-    &__triangle {
-      display: none;
+type InputWrapperProps = {
+  isError?: boolean;
+  isOpen: boolean;
+  placeTop: boolean;
+};
+export const InputWrapper = styled.div<InputWrapperProps>`
+  ${(props: InputWrapperProps) => css`
+    border-radius: 0.75rem;
+    box-sizing: border-box;
+    display: flex;
+    align-items: center;
+    gap: 0.25rem;
+    background-color: var(--input-bg);
+    min-height: var(--input-min-height);
+    padding-right: 2.5rem;
+    padding-left: var(--input-left-padding);
+    margin: 0;
+    border-width: 1px;
+    border-style: solid;
+    border-color: var(--input-border-color);
+    transition: box-shadow 0.1s ease-out;
+    ${props.isOpen &&
+    css`
+      box-shadow: var(--input-box-shadow);
+    `}
+    ${props.isError &&
+    css`
+      border-color: var(--input-border-color-error);
+      box-shadow: var(--input-box-shadow-error);
+    `}
+    ${props.isOpen &&
+    !props.placeTop &&
+    css`
+      border-bottom-left-radius: 0;
+      border-bottom-right-radius: 0;
+      border-bottom-color: transparent;
+    `}
+    ${props.isOpen &&
+    props.placeTop &&
+    css`
+      border-top-left-radius: 0;
+      border-top-right-radius: 0;
+      border-top-color: transparent;
+    `}
+> svg {
+      color: ${colors.fc_black_100};
+      position: absolute;
+      right: 0.5rem;
+      bottom: 0.5rem;
+      height: var(--input-icon-height);
+      width: auto;
     }
+  `}
+`;
 
-    &__day-names {
-      display: flex;
-      justify-content: space-between;
-      height: 2rem;
-    }
-
-    &__day-name {
-      color: ${farmerConnectTheme.colors.fc_grey};
-      text-align: center;
-      font-size: 0.6875rem;
-      flex: 1;
-      line-height: 2rem;
-      font-weight: 500;
-    }
-
-    &__month {
-      width: 16rem;
-    }
-
-    &__week {
-      display: flex;
-      justify-content: space-between;
-      height: 2rem;
-
-      & + .react-datepicker__week {
-        margin-top: 0.25rem;
-      }
-    }
-
-    &__day {
-      flex: 1;
-      display: inline-flex;
-      align-items: center;
-      justify-content: center;
-      font-size: 0.6875rem;
-      color: ${farmerConnectTheme.colors.fc_black_100};
+export const DateInput = styled.input`
+  ${(props) => css`
+    font-family: 'Red Hat Text', sans-serif;
+    box-sizing: border-box;
+    display: inline-flex;
+    margin: 0;
+    padding: 0;
+    flex: 1;
+    font-size: var(--input-font-size);
+    font-weight: var(--input-font-weight);
+    outline: none;
+    border: none;
+    text-transform: uppercase;
+    padding: 0.5rem 0;
+    background-color: transparent;
+    &::placeholder {
+      color: ${colors.fc_black_30};
+      text-transform: none;
       font-weight: 400;
-      cursor: pointer;
-      position: relative;
-
-      &:hover {
-        > div {
-          background-color: ${farmerConnectTheme.colors.fc_light_green};
-        }
-      }
-
-      &--outside-month {
-        color: ${farmerConnectTheme.colors.fc_black_30};
-      }
-
-      &--selected {
-        position: relative;
-        font-weight: 700;
-        color: ${farmerConnectTheme.colors.fc_white};
-
-        > div {
-          background-color: ${farmerConnectTheme.colors.fc_green};
-        }
-      }
-
-      &--in-range {
-        background: ${farmerConnectTheme.colors.fc_black_5};
-      }
-
-      &--in-selecting-range {
-        background: ${farmerConnectTheme.colors.fc_black_5};
-      }
-
-      &--range-start {
-        border-radius: 2rem 0 0 2rem;
-        color: ${farmerConnectTheme.colors.fc_white};
-        font-weight: 700;
-        position: relative;
-
-        &::before {
-          content: '';
-          position: absolute;
-          inset: 0 50% 0 0;
-          background-color: ${farmerConnectTheme.colors.fc_white};
-        }
-
-        > div {
-          background-color: ${farmerConnectTheme.colors.fc_green};
-          z-index: 1;
-        }
-      }
-
-      &--range-end {
-        border-radius: 0 2rem 2rem 0;
-        color: ${farmerConnectTheme.colors.fc_white};
-        font-weight: 700;
-
-        &::before {
-          content: '';
-          position: absolute;
-          inset: 0 0 0 50%;
-          background-color: ${farmerConnectTheme.colors.fc_white};
-        }
-
-        > div {
-          background-color: ${farmerConnectTheme.colors.fc_green};
-          z-index: 1;
-        }
-      }
-
-      &--range-start.react-datepicker__day--range-end {
-        background-color: ${farmerConnectTheme.colors.fc_white};
-
-        &::before {
-          display: none;
-        }
-      }
-
-      &--selecting-range-start,
-      &--selecting-range-end {
-        border-radius: 2rem 0 0 2rem;
-        background-color: ${farmerConnectTheme.colors.fc_black_5};
-        color: ${farmerConnectTheme.colors.fc_white};
-        font-weight: 700;
-
-        &::before {
-          content: '';
-          position: absolute;
-          inset: 0 50% 0 0;
-          background-color: ${farmerConnectTheme.colors.fc_white};
-        }
-
-        > div {
-          background-color: ${farmerConnectTheme.colors.fc_green};
-          z-index: 1;
-        }
-      }
-
-      &--in-selecting-range:hover {
-        &::before {
-          content: '';
-          position: absolute;
-          inset: 0 0 0 50%;
-          background-color: ${farmerConnectTheme.colors.fc_white};
-        }
-
-        > div {
-          background-color: ${farmerConnectTheme.colors.fc_light_green};
-          z-index: 1;
-        }
-      }
-
-      &--excluded {
-        color: ${farmerConnectTheme.colors.fc_black_30};
-        background-color: ${farmerConnectTheme.colors.fc_white};
-
-        &::after {
-          content: '';
-          position: absolute;
-          height: 1.125rem;
-          width: 0.125rem;
-          background-color: ${farmerConnectTheme.colors.fc_black_30};
-          top: 50%;
-          left: 50%;
-          transform: translate(-50%, -50%) rotate(45deg);
-        }
-      }
     }
-  }
+    &:focus {
+      font-weight: 700;
+    }
+    ::-webkit-inner-spin-button,
+    ::-webkit-outer-spin-button {
+      -webkit-appearance: none;
+      background-color: black;
+    }
+    &.align-right {
+      text-align: right;
+    }
+  `}
 `;
 
-export const Day = styled.div`
-  height: 2rem;
-  width: 2rem;
+export const DateInputDivider = styled(Typography).attrs({
+  variant: 'small',
+  fontStyle: 'highlight',
+  tagName: 'span',
+})`
   text-align: center;
-  line-height: 2rem;
-  margin: 0;
-  border-radius: 1rem;
+  font-size: var(--input-font-size);
 `;
 
-export const BottomRow = styled.div`
-  grid-column: 1/-1;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  margin-top: 1rem;
-  flex: 1;
-  padding: 0 0.5rem;
+type CalendarWrapperProps = {
+  placeTop: boolean;
+  isError?: boolean;
+};
+export const CalendarWrapper = styled.div<CalendarWrapperProps>`
+  ${(props) => css`
+    border-radius: 0.75rem;
+    border-width: 1px;
+    border-style: solid;
+    border-color: var(--input-border-color);
+    box-shadow: var(--input-box-shadow);
+    background-color: var(--input-bg);
+    margin: -1px 0;
+    z-index: 100;
+    ${props.isError &&
+    css`
+      border-color: var(--input-border-color-error);
+      box-shadow: var(--input-box-shadow-error);
+    `}
+    ${props.placeTop
+      ? css`
+          border-bottom-left-radius: 0;
+        `
+      : css`
+          border-top-left-radius: 0;
+        `}
+  `}
 `;
 
-export const CancelButton = styled(CustomButton)``;
+type CalendarContainerProps = {
+  inputWidth: number;
+  isOpen: boolean;
+  placeTop: boolean;
+};
+export const CalendarContainer = styled.div<CalendarContainerProps>`
+  ${(props) => css`
+    position: relative;
+    padding: 1.5rem;
+    &:after {
+      content: '';
+      position: absolute;
+      height: 0.25rem;
+      width: ${props.inputWidth}px;
+      background-color: var(--mask-divider-color);
+      ${props.isOpen &&
+      props.placeTop &&
+      css`
+        bottom: -4px;
+      `};
+      ${props.isOpen &&
+      !props.placeTop &&
+      css`
+        top: -4px;
+      `};
+      left: 0;
+    }
+  `}
+`;
 
 export const Header = styled.div`
-  position: relative;
-
-  > h4 {
-    font-size: 0.875rem;
-    font-weight: 700;
-    color: ${farmerConnectTheme.colors.fc_black_100};
-    margin: 0 0 1rem 0;
-    text-align: center;
-  }
-
-  > div {
-    display: flex;
-    gap: 0.5rem;
-
-    .select__control {
-      flex: 1;
-    }
-  }
-`;
-
-type ArrowButtonProps = {
-  right?: boolean;
-};
-
-export const ArrowButton = styled(CustomButton)<ArrowButtonProps>`
-  position: absolute;
-  inset: ${({ right }) => (right ? '-0.25rem 0 auto auto' : '-0.25rem auto auto 0')};
-  color: ${farmerConnectTheme.colors.fc_black_70};
-  height: 2rem;
-  width: 2rem;
-  padding: 0.5rem;
-`;
-
-export const ChangeYearButton = styled.div`
-  position: absolute;
-  left: calc(50% - 4.5rem);
-  top: 1.5rem;
-  background-color: #f3f3f3;
-  border-radius: 0.75rem;
-  width: 9rem;
   display: flex;
   justify-content: space-between;
+  > div:first-child {
+    width: 8.25rem;
+  }
+`;
+
+export const DefaultYearSelect = styled(CustomSelect)``;
+export const SmallYearSelect = styled(SmallSelect)``;
+
+export const ButtonGroup = styled.div`
+  display: flex;
+  gap: 0.5rem;
+`;
+
+export const CalendarBody = styled.div`
+  position: relative;
+  margin-top: 1.5rem;
+  display: flex;
+  gap: 1rem;
+`;
+
+export const SingleCalendar = styled.div`
+  height: 18rem;
+  display: grid;
+  grid-template-columns: repeat(7, 2.5rem);
+  grid-auto-rows: 2.25rem;
   align-items: center;
-  > button {
-    border: none;
-    outline: none;
-    height: 2rem;
-    width: 2rem;
-    background-color: transparent;
-    cursor: pointer;
-    display: inline-flex;
+  justify-content: center;
+`;
+
+export const MonthName = styled(Typography).attrs({
+  fontStyle: 'highlight',
+})`
+  text-align: center;
+  grid-column: 1/8;
+  pointer-events: none;
+`;
+
+type DayProps = {
+  isSelected: boolean;
+  isInRange: boolean;
+  isInSelectingRange: boolean;
+  isHovering: boolean;
+  isStartDate: boolean;
+  isEndDate: boolean;
+};
+export const Day = styled.button<DayProps>`
+  ${(props) => css`
+    position: relative;
+    padding: 0;
+    margin: 0;
+    display: flex;
     align-items: center;
     justify-content: center;
-    transition: color 0.1s ease-out;
-    padding: 0;
-    color: #b9b9b9;
-    &:hover {
-      color: #5b5b5b;
+    font-family: 'Red Hat Text', sans-serif;
+    font-size: 0.6875rem;
+    color: ${colors.fc_black_100};
+    height: 2rem;
+    width: 2.5rem;
+    cursor: pointer;
+    border: none;
+    background-color: transparent;
+    transition: all 0.1s ease-out;
+    > span {
+      height: 2rem;
+      width: 2rem;
+      flex-shrink: 0;
+      border-radius: 1rem;
+      display: flex;
+      align-items: center;
+      justify-content: center;
+      transition: background-color 0.1s ease-out;
     }
-  }
-  > span {
-    font-size: 0.875rem;
-    color: #141414;
-    font-weight: 700;
-    cursor: default;
-  }
+    ${props.isStartDate &&
+    css`
+      box-shadow: -1.125rem 0 0 0 var(--range-bg-color) inset;
+    `}
+    ${props.isEndDate &&
+    css`
+      box-shadow: 1.125rem 0 0 0 var(--range-bg-color) inset;
+    `}
+    ${props.isStartDate &&
+    props.isEndDate &&
+    css`
+      box-shadow: none !important;
+    `}
+    ${props.isInRange &&
+    css`
+      background-color: var(--range-bg-color);
+      color: var(--range-font-color);
+    `}
+      ${props.isInSelectingRange &&
+    css`
+      background-color: var(--range-bg-color);
+      color: var(--range-font-color);
+    `}
+    ${props.isHovering &&
+    css`
+      > span {
+        background-color: ${colors.fc_light_green};
+      }
+    `}
+    ${props.isSelected &&
+    css`
+      > span {
+        background-color: ${colors.fc_green};
+        font-weight: 700;
+        color: ${colors.fc_white};
+      }
+    `}
+    ${props.isHovering &&
+    props.isSelected &&
+    css`
+      > span {
+        background-color: ${colors.fc_hover_green};
+      }
+    `}
+  `}
+`;
+
+type MonthNavigationButtonProps = {
+  align: 'left' | 'right';
+};
+export const MonthNavigationButton = styled(CustomButton)<MonthNavigationButtonProps>`
+  border: none;
+  position: absolute;
+  top: 0.5rem;
+  height: 1.5rem;
+  width: 1.5rem;
+  padding: 0;
+  min-width: auto;
+  min-height: auto;
+  ${(props) =>
+    props.align === 'left' &&
+    css`
+      left: 0;
+    `}
+  ${(props) =>
+    props.align === 'right' &&
+    css`
+      right: 0;
+    `}
+`;
+
+export const CalendarIcon = styled(Calendar)``;
+
+export const WeekdayName = styled(Typography).attrs({
+  tagName: 'span',
+  fontStyle: 'semibold',
+  variant: 'small',
+})`
+  text-align: center;
+  pointer-events: none;
 `;
